Fixed some bugs and passed all the testbenches so far.

+++ app/controllers/favorites_controller.rb
@@ -3,6 +3,8 @@ class FavoritesController < ApplicationController
     @post = Post.find(params[:post_id])
     favorite = current_user.favorites.build(post: @post)
 
+    authorize favorite
+
     if favorite.save
       flash[:notice] = "You've favorited this post."
       redirect_to [@post.topic, @post]
@@ -14,7 +16,9 @@ class FavoritesController < ApplicationController
 
   def destroy
     @post = Post.find(params[:post_id])
-    favorite = current_user.favorites.where(post_id: @post.id).first
+    favorite = current_user.favorites.find(params[:post_id])
+
+    authorize favorite
 
     if favorite.destroy
       flash[:notice] = "You've unfavorited this post."
